QRStorage complete Will delete qr codes afer a fixed limit of time, and if there is to many in storage
