+++ app/src/main/java/no/ntnu/tpg4850/frugalar/scanner/QRScanner.java
@@ -19,7 +19,7 @@ public class QRScanner implements Camera.PreviewCallback {
     private ImageScanner mScanner;
     private CardboardOverlayView mOverlayView;
     private Camera mCamera;
-    //TODO:Handle more than 1 qr at the same time.
+    private QRStorage storage;
     //TODO: Interpolation between detections, and trail off behavior. IE. QR object with bounds and
     //TODO:id kept for a fixed period of time, and removed if not renewed by a new sighting. Can also use accelorometer to improve this behavior.
     private int[] qrCodeBounds = null;
@@ -27,6 +27,7 @@ public class QRScanner implements Camera.PreviewCallback {
 
     public QRScanner(CardboardOverlayView mOverlayView) {
         this.mOverlayView = mOverlayView; //TODO: TEMP WAY of showing data in text hud
+        storage = new QRStorage(10, 1000);
         setupScanner();
     }
 
@@ -45,32 +46,9 @@ public class QRScanner implements Camera.PreviewCallback {
         }
     }
 
-    public Point getMidpoint(int id) {
-        //TODO: Only 1 qr code can be kept so id irrelevant. Use 0 as value.
-        int[] v = this.qrCodeBounds;
-        if(v != null && v.length>=3) {
-            Point p = new Point((v[0]+v[2])/2, (v[1]+v[3])/2);
-            return p;
-        }
-        return null;
-    }
 
-    public Point[] getBounds(int id) {
-        int[] v = this.qrCodeBounds;
-        int corners = 4;
-        Point[] p = new Point[corners];
-        if(v != null && v.length>=3) {
-            p[0] = new Point(v[0], v[1]);
-            p[1] = new Point(v[0]+v[2], v[1]);
-            p[2] = new Point(v[0]+v[2], v[1]+ v[3]);
-            p[3] = new Point(v[0], v[1]+ v[3]);
-        }
-        return p;
-    }
 
-    public int[] getBoundsRect(int id) {
-        return this.qrCodeBounds;
-    }
+
 
     @Override
     public void onPreviewFrame(byte[] data, Camera camera) {
@@ -87,16 +65,15 @@ public class QRScanner implements Camera.PreviewCallback {
             SymbolSet syms = mScanner.getResults();
             Log.i(TAG, syms.toString());
             for (Symbol sym : syms) {
-                //TODO: How to handle more than 1 qr in the image.
-                //Currently assume that only one qr code is detected for each preview frame.
                 Log.i(TAG, "barcode result " + sym.getData());
-                this.qrId = sym.getData();
-                this.qrCodeBounds = sym.getBounds();
-                for(int i = 0; i<this.qrCodeBounds.length; i++) {
-                    Log.i(TAG, this.qrCodeBounds[i] + "");
-                }
+                QRCode q = new QRCode(sym.getData(), sym.getBounds());
+                Log.i(TAG, this.storage.size()+ "");
+                this.storage.Store(q);
+                //TODO: async storage update every so and so ms
+
             }
         }
+        this.storage.updateAll();
 
     }
 }
